fix(frontend): resolve exhaustive-deps and fast-refresh lint warnings
