+++ frontend/src/components/PerformanceGraph.tsx
@@ -12,6 +12,7 @@ import {
 } from 'recharts';
 import type { WorkoutSet } from '../types';
 import { EXERCISES } from '../routines';
+import { calculateEstimated1RM } from '../utils';
 
 interface PerformanceGraphProps {
     history: WorkoutSet[];
@@ -19,16 +20,6 @@ interface PerformanceGraphProps {
     theme: 'dark' | 'light';
 }
 
-/**
- * Brzycki式を用いた推定1RMの計算
- * 1RM = Weight * (36 / (37 - Reps))
- * ※Repsが10回以下で精度が高いとされる
- */
-export const calculateEstimated1RM = (weight: number, reps: number): number => {
-    if (reps <= 0) return 0;
-    if (reps === 1) return weight;
-    return weight * (36 / (37 - reps));
-};
 
 export default function PerformanceGraph({ history, exerciseId, theme }: PerformanceGraphProps) {
     // グラフデータの生成
